Removed Threading from Video File Processing & Added README.md Template
A README.md template has been added. Everyone in the 2XStudy team, use this template to construct the final README.md file for the master branch.

A Note About Threading:
After a number of thorough testing, threading does NOT improve the runtime of video generation. In fact, if anything, threading seems to SLOW down the operation by a bit.